backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "csv-transformer api"
authors = ["TomJan533 <tomaszjan533@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.8"
django-cors-headers = "^4.4.0"
python-decouple = "^3.8"
requests = "^2.32.3"
kafka-python = "^2.0.2"
channels = "^4.1.0"
channels-redis = "^4.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"
mock = "^5.1.0"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pre-commit = "^3.8.0"
pytest-dotenv = "^0.5.2"
pytest-mock-resources = "^2.12.0"
python-on-whales = "^0.72.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
skip = ["api/migrations"]